+++ R/batchtools_local.R
@@ -7,9 +7,9 @@
 #' be assigned to the calling environment rather than to a local one).
 #' Both types of futures will block until the futures are resolved.
 #'
-#' @inheritParams BatchtoolsUniprocessFuture
+#' @inheritParams BatchtoolsFutureBackend
 #' 
-#' @param \ldots Additional arguments passed to [BatchtoolsUniprocessFuture()].
+#' @param \ldots Additional arguments passed to [BatchtoolsFutureBackend()].
 #'
 #' @return An object of class `BatchtoolsUniprocessFuture`.
 #'
@@ -30,14 +30,29 @@
 #'
 #' @example incl/batchtools_local.R
 #'
+#' @rdname BatchtoolsInteractiveFutureBackend
+#'
 #' @importFrom batchtools makeClusterFunctionsInteractive
-#' @aliases batchtools_interactive batchtools_bash
+#' @aliases batchtools_local batchtools_interactive batchtools_bash
+#' @aliases BatchtoolsLocalFutureBackend BatchtoolsInteractiveFutureBackend BatchtoolsBashFutureBackend
+#' @export
+BatchtoolsLocalFutureBackend <- function(...) {
+  assert_no_positional_args_but_first()
+
+  core <- BatchtoolsUniprocessFutureBackend(
+    cluster.functions = makeClusterFunctionsInteractive(external = TRUE),
+    ...
+  )
+
+  core[["futureClasses"]] <- c("BatchtoolsLocalFuture", "BatchtoolsUniprocessFuture", core[["futureClasses"]])
+  core <- structure(core, class = c("BatchtoolsLocalFutureBackend", class(core)))
+  core
+}
+
+
 #' @export
 batchtools_local <- function(..., envir = parent.frame()) {
-  cf <- makeClusterFunctionsInteractive(external = TRUE)
-  future <- BatchtoolsLocalFuture(..., envir = envir, cluster.functions = cf)
-  if (!future$lazy) future <- run(future)
-  invisible(future)
+ stop("INTERNAL ERROR: The future.batchtools::batchtools_local() must never be called directly")
 }
 class(batchtools_local) <- c(
   "batchtools_local", "batchtools_uniprocess", "batchtools",
@@ -45,3 +60,8 @@ class(batchtools_local) <- c(
 )
 attr(batchtools_local, "tweakable") <- c("finalize")
 attr(batchtools_local, "untweakable") <- c("workers")
+attr(batchtools_local, "init") <- TRUE
+attr(batchtools_local, "factory") <- BatchtoolsLocalFutureBackend
+
+
+